Implemented logic for removing a player. Fleshed out logic for adding a player.
There is a lot of extra logic that goes into the financial side of
things. So I deferred quite a bit of that logic to future service
classes that will (hopefully) be able to handle it all.

Future functionality to implement within the team class: moving a player
from one roster to another, separate active roster from bench, check
validity of starting lineup, display rosters.