Started implementing League management class.
A few things I'm considering are mentioned in the TODO's of the file.
Most notably, simplifying certain processes whenever I incorporate the
Sleeper API for handling player performances; also implementing some
form of a state machine for handling the League's calendar.

There are a lot of moving parts to La Liga, so having something that can
help manage when certain actions are possible and what phase of the
season we're in would be a rather large help.