Added some basic functionality for the CLI.
The Click library is extremely useful, still figuring out the kinks to
it. I will need a way of saving the game state before this functionality
can really start to be useful. Right now you are able to set up a demo
league, display the league status, display a specific team's roster
(which will be more useful if I can actually save and search for a
team), advance the league's season, and extend a player's contract.

More things that must be added to this basic CLI is checking for
holdouts and, of course, simulating a draft. I also should make more
decorative console output messages. Once this is done, it will also be
helpful to create a package entry point for the application to
immediately be able to call the program's commands.