Added a holdout check procedure and a stub for the rookie draft.
Lots to do, but this CLI will help me with testing out the system a
little more. With future functionality I can begin adding new features
to the CLI in order to test them (and to start building up the product).

This CLI is mostly focused on league functionality (some basic
overarching things for the League commissioner/manager to worry about).
But of course, this will soon become more focused on individual team and
roster management.