Implemented test suites for contracts and players.
First time working with pytest, definitely need to read up more on the
documentation. But it makes creating and running these test suites very
simple and easy to implement.

Next I will need to implement the test classes for the league and the
team models. During this time I should refresh myself on proper testing
frameworks and practices.